requirements.txt
asgiref==3.8.1
attrs==24.2.0
coverage==7.6.1
Django==5.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==28.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
